Fix compilation from spcomp 7032+
The sourcemod compiler changed some things from 7032 and on that breaks
this macro if there is a space after the comma.

It appears to not be a regression and won't be fixed, so we can just
remove the comma and deal with it.
See: https://github.com/alliedmodders/sourcepawn/issues/879